added light/dark toggle to map in about section to a seperate component, did some asset reorganization and imports updated
