+++ components/RotatingBookCovers.tsx
@@ -3,14 +3,14 @@ import { motion, AnimatePresence } from 'framer-motion'
 import Image from 'next/image'
 import Card from './Card'
 import CardHeader from './CardHeader'
-import AlanBookCover from '@/assets/Book7-AlanWatts.jpeg'
-import CosmosBookCover from '@/assets/Book6-Cosmos.jpg'
-import NatGeoBookCover from '@/assets/Book2-natgeo.jpg'
-import PoohBookCover from '@/assets/Book4-Pooh.jpg'
-import RiedelBookCover from '@/assets/Book3-AnorganischeChemie.jpg'
-import HarryPotterBookCover from '@/assets/Book5-HarryPotter.jpg'
-import PragmaticProgrammerBookCover from '@/assets/Book1-PragmaticProgrammer.jpg'
-import LifeOfPiBookCover from '@/assets/Book8-LifeOfPi.jpg'
+import AlanBookCover from '@/assets/books/Book7-AlanWatts.jpeg'
+import CosmosBookCover from '@/assets/books/Book6-Cosmos.jpg'
+import NatGeoBookCover from '@/assets/books/Book2-natgeo.jpg'
+import PoohBookCover from '@/assets/books/Book4-Pooh.jpg'
+import RiedelBookCover from '@/assets/books/Book3-AnorganischeChemie.jpg'
+import HarryPotterBookCover from '@/assets/books/Book5-HarryPotter.jpg'
+import PragmaticProgrammerBookCover from '@/assets/books/Book1-PragmaticProgrammer.jpg'
+import LifeOfPiBookCover from '@/assets/books/Book8-LifeOfPi.jpg'
 
 // this was good ol friend, he said its more lightweight and ssr friendly than using "swiper" which i know.
